+++ react-components/src/components/RevealToolbar/LayersContainer/WholeLayerVisibilityToggle.tsx
@@ -4,7 +4,7 @@
 import { Checkbox, Flex } from '@cognite/cogs.js';
 import { type ModelHandler } from './ModelHandler';
 import { type ChangeEvent, useCallback, useMemo, type ReactElement, type MouseEvent } from 'react';
-import { StyledChipCount, StyledLabel } from '../../RevealToolbar/LayersContainer/elements';
+import { StyledChipCount, StyledLabel } from './elements';
 import { type UpdateModelHandlersCallback } from './useModelHandlers';
 import { useReveal } from '../../RevealCanvas/ViewerContext';
