+++ react-components/src/components/RevealToolbar/LayersContainer/updateExternalStateFromLayerHandlers.tsx
